+++ tests/test36_numpy_videograin.py
@@ -1,0 +1,574 @@
+#!/usr/bin/python
+#
+# Copyright 2018 British Broadcasting Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+from unittest import TestCase
+
+import uuid
+from mediagrains.numpy import VideoGrain
+from mediagrains.cogenums import (
+    CogFrameFormat,
+    CogFrameLayout,
+    COG_FRAME_FORMAT_BYTES_PER_VALUE,
+    COG_FRAME_FORMAT_H_SHIFT,
+    COG_FRAME_FORMAT_V_SHIFT,
+    COG_FRAME_IS_PACKED,
+    COG_FRAME_IS_COMPRESSED,
+    COG_FRAME_IS_PLANAR,
+    COG_FRAME_IS_PLANAR_RGB,
+    COG_FRAME_FORMAT_ACTIVE_BITS)
+from mediatimestamp.immutable import Timestamp, TimeRange
+import mock
+from fractions import Fraction
+from copy import copy, deepcopy
+
+import numpy as np
+
+
+class TestGrain (TestCase):
+    def _get_bitdepth(self, fmt):
+        if COG_FRAME_IS_PLANAR(fmt):
+            return COG_FRAME_FORMAT_ACTIVE_BITS(fmt)
+        elif fmt in [CogFrameFormat.UYVY,
+                     CogFrameFormat.YUYV,
+                     CogFrameFormat.RGB,
+                     CogFrameFormat.RGBx,
+                     CogFrameFormat.RGBA,
+                     CogFrameFormat.BGRx,
+                     CogFrameFormat.BGRx,
+                     CogFrameFormat.ARGB,
+                     CogFrameFormat.xRGB,
+                     CogFrameFormat.ABGR,
+                     CogFrameFormat.xBGR]:
+            return 8
+        elif fmt == CogFrameFormat.v216:
+            return 16
+        elif fmt == CogFrameFormat.v210:
+            return 10
+        else:
+            raise Exception()
+
+    def _get_hs_vs_and_bps(self, fmt):
+        if COG_FRAME_IS_PLANAR(fmt):
+            return (COG_FRAME_FORMAT_H_SHIFT(fmt), COG_FRAME_FORMAT_V_SHIFT(fmt), COG_FRAME_FORMAT_BYTES_PER_VALUE(fmt))
+        elif fmt in [CogFrameFormat.UYVY, CogFrameFormat.YUYV]:
+            return (1, 0, 1)
+        elif fmt in [CogFrameFormat.RGB,
+                     CogFrameFormat.RGBx,
+                     CogFrameFormat.RGBA,
+                     CogFrameFormat.BGRx,
+                     CogFrameFormat.BGRx,
+                     CogFrameFormat.ARGB,
+                     CogFrameFormat.xRGB,
+                     CogFrameFormat.ABGR,
+                     CogFrameFormat.xBGR]:
+            return (0, 0, 1)
+        elif fmt == CogFrameFormat.v216:
+            return (1, 0, 2)
+        elif fmt == CogFrameFormat.v210:
+            return (1, 0, 4)
+        else:
+            raise Exception()
+
+    def _is_rgb(self, fmt):
+        if COG_FRAME_IS_PLANAR(fmt):
+            return COG_FRAME_IS_PLANAR_RGB(fmt)
+        elif fmt in [CogFrameFormat.UYVY,
+                     CogFrameFormat.YUYV,
+                     CogFrameFormat.v216,
+                     CogFrameFormat.v210]:
+            return False
+        elif fmt in [CogFrameFormat.RGB,
+                     CogFrameFormat.RGBx,
+                     CogFrameFormat.RGBA,
+                     CogFrameFormat.BGRx,
+                     CogFrameFormat.BGRx,
+                     CogFrameFormat.ARGB,
+                     CogFrameFormat.xRGB,
+                     CogFrameFormat.ABGR,
+                     CogFrameFormat.xBGR]:
+            return True
+        else:
+            raise Exception()
+
+    def assertComponentsAreModifiable(self, grain):
+        width = grain.width
+        height = grain.height
+        fmt = grain.format
+
+        (hs, vs, _) = self._get_hs_vs_and_bps(fmt)
+
+        # Test that changes to the component arrays are reflected in the main data array
+        for y in range(0, 16):
+            for x in range(0, 16):
+                grain.component_data[0][x, y] = (y*16 + x) & 0x3F
+
+        for y in range(0, 16 >> vs):
+            for x in range(0, 16 >> hs):
+                grain.component_data[1][x, y] = (y*16 + x) & 0x3F + 0x40
+                grain.component_data[2][x, y] = (y*16 + x) & 0x3F + 0x50
+
+        if COG_FRAME_IS_PLANAR(fmt):
+            for y in range(0, 16):
+                for x in range(0, 16):
+                    self.assertEqual(grain.data[y*width + x], (y*16 + x) & 0x3F)
+
+            for y in range(0, 16 >> vs):
+                for x in range(0, 16 >> hs):
+                    self.assertEqual(grain.data[width*height + y*(width >> hs) + x], (y*16 + x) & 0x3F + 0x40)
+                    self.assertEqual(grain.data[width*height + (width >> hs)*(height >> vs) + y*(width >> hs) + x], (y*16 + x) & 0x3F + 0x50)
+
+        elif fmt in [CogFrameFormat.UYVY, CogFrameFormat.v216]:
+            for y in range(0, 16):
+                for x in range(0, 8):
+                    self.assertEqual(grain.data[y*width*2 + 4*x + 0], (y*16 + x) & 0x3F + 0x40)
+                    self.assertEqual(grain.data[y*width*2 + 4*x + 1], (y*16 + 2*x + 0) & 0x3F)
+                    self.assertEqual(grain.data[y*width*2 + 4*x + 2], (y*16 + x) & 0x3F + 0x50)
+                    self.assertEqual(grain.data[y*width*2 + 4*x + 3], (y*16 + 2*x + 1) & 0x3F)
+
+        elif fmt == CogFrameFormat.YUYV:
+            for y in range(0, 16):
+                for x in range(0, 8):
+                    self.assertEqual(grain.data[y*width*2 + 4*x + 0], (y*16 + 2*x + 0) & 0x3F)
+                    self.assertEqual(grain.data[y*width*2 + 4*x + 1], (y*16 + x) & 0x3F + 0x40)
+                    self.assertEqual(grain.data[y*width*2 + 4*x + 2], (y*16 + 2*x + 1) & 0x3F)
+                    self.assertEqual(grain.data[y*width*2 + 4*x + 3], (y*16 + x) & 0x3F + 0x50)
+
+        elif fmt == CogFrameFormat.RGB:
+            for y in range(0, 16):
+                for x in range(0, 16):
+                    self.assertEqual(grain.data[y*width*3 + 3*x + 0], (y*16 + x) & 0x3F)
+                    self.assertEqual(grain.data[y*width*3 + 3*x + 1], (y*16 + x) & 0x3F + 0x40)
+                    self.assertEqual(grain.data[y*width*3 + 3*x + 2], (y*16 + x) & 0x3F + 0x50)
+
+        elif fmt in [CogFrameFormat.RGBx,
+                     CogFrameFormat.RGBA,
+                     CogFrameFormat.BGRx,
+                     CogFrameFormat.BGRx]:
+            for y in range(0, 16):
+                for x in range(0, 16):
+                    self.assertEqual(grain.data[y*width*4 + 4*x + 0], (y*16 + x) & 0x3F)
+                    self.assertEqual(grain.data[y*width*4 + 4*x + 1], (y*16 + x) & 0x3F + 0x40)
+                    self.assertEqual(grain.data[y*width*4 + 4*x + 2], (y*16 + x) & 0x3F + 0x50)
+
+        elif fmt in [CogFrameFormat.ARGB,
+                     CogFrameFormat.xRGB,
+                     CogFrameFormat.ABGR,
+                     CogFrameFormat.xBGR]:
+            for y in range(0, 16):
+                for x in range(0, 16):
+                    self.assertEqual(grain.data[y*width*4 + 4*x + 1], (y*16 + x) & 0x3F)
+                    self.assertEqual(grain.data[y*width*4 + 4*x + 2], (y*16 + x) & 0x3F + 0x40)
+                    self.assertEqual(grain.data[y*width*4 + 4*x + 3], (y*16 + x) & 0x3F + 0x50)
+
+        else:
+            raise Exception()
+
+    def assertIsVideoGrain(self,
+                           fmt,
+                           src_id=uuid.UUID("f18ee944-0841-11e8-b0b0-17cef04bd429"),
+                           flow_id=uuid.UUID("f79ce4da-0841-11e8-9a5b-dfedb11bafeb"),
+                           ots=Timestamp.from_tai_sec_nsec("417798915:5"),
+                           sts=Timestamp.from_tai_sec_nsec("417798915:10"),
+                           cts=Timestamp.from_tai_sec_nsec("417798915:0"),
+                           rate=Fraction(25, 1),
+                           width=1920,
+                           height=1080):
+        def __inner(grain):
+            self.assertEqual(grain.grain_type, "video")
+            self.assertEqual(grain.source_id, src_id)
+            self.assertEqual(grain.flow_id, flow_id)
+            self.assertEqual(grain.origin_timestamp, ots)
+            self.assertEqual(grain.final_origin_timestamp(), ots)
+            self.assertEqual(grain.origin_timerange(), TimeRange.from_single_timestamp(ots))
+            self.assertEqual(grain.sync_timestamp, sts)
+            self.assertEqual(grain.creation_timestamp, cts)
+            self.assertEqual(grain.rate, rate)
+            self.assertEqual(grain.duration, 1/rate)
+            self.assertEqual(grain.timelabels, [])
+            self.assertEqual(grain.format, fmt)
+            self.assertEqual(grain.width, width)
+            self.assertEqual(grain.height, height)
+            self.assertEqual(grain.layout, CogFrameLayout.FULL_FRAME)
+            self.assertEqual(grain.extension, 0)
+            self.assertIsNone(grain.source_aspect_ratio)
+            self.assertIsNone(grain.pixel_aspect_ratio)
+
+            (hs, vs, bps) = self._get_hs_vs_and_bps(fmt)
+
+            if COG_FRAME_IS_PLANAR(fmt):
+                self.assertEqual(len(grain.components), 3)
+                self.assertEqual(grain.components[0].stride, width*bps)
+                self.assertEqual(grain.components[0].width, width)
+                self.assertEqual(grain.components[0].height, height)
+                self.assertEqual(grain.components[0].offset, 0)
+                self.assertEqual(grain.components[0].length, width*height*bps)
+                self.assertEqual(len(grain.components[0]), 5)
+
+                self.assertEqual(grain.components[1].stride, width*bps >> hs)
+                self.assertEqual(grain.components[1].width, width >> hs)
+                self.assertEqual(grain.components[1].height, height >> vs)
+                self.assertEqual(grain.components[1].offset, width*height*bps)
+                self.assertEqual(grain.components[1].length, width*height*bps >> (hs + vs))
+                self.assertEqual(len(grain.components[1]), 5)
+
+                self.assertEqual(grain.components[2].stride, width*bps >> hs)
+                self.assertEqual(grain.components[2].width, width >> hs)
+                self.assertEqual(grain.components[2].height, height >> vs)
+                self.assertEqual(grain.components[2].offset, width*height*bps + (width*height*bps >> (hs + vs)))
+                self.assertEqual(grain.components[2].length, width*height*bps >> (hs + vs))
+                self.assertEqual(len(grain.components[2]), 5)
+
+                self.assertEqual(grain.expected_length, (width*height + 2*(width >> hs)*(height >> vs))*bps)
+            elif fmt in [CogFrameFormat.UYVY, CogFrameFormat.YUYV]:
+                self.assertEqual(len(grain.components), 1)
+                self.assertEqual(grain.components[0].stride, width*bps + 2*(width >> hs)*bps)
+                self.assertEqual(grain.components[0].width, width)
+                self.assertEqual(grain.components[0].height, height)
+                self.assertEqual(grain.components[0].offset, 0)
+                self.assertEqual(grain.components[0].length, width*height*bps*2)
+                self.assertEqual(len(grain.components[0]), 5)
+
+                self.assertEqual(grain.expected_length, width*height*bps*2)
+            elif fmt in [CogFrameFormat.RGB]:
+                self.assertEqual(len(grain.components), 1)
+                self.assertEqual(grain.components[0].stride, 3*width*bps)
+                self.assertEqual(grain.components[0].width, width)
+                self.assertEqual(grain.components[0].height, height)
+                self.assertEqual(grain.components[0].offset, 0)
+                self.assertEqual(grain.components[0].length, width*height*bps*3)
+                self.assertEqual(len(grain.components[0]), 5)
+            elif fmt in [CogFrameFormat.RGBx,
+                         CogFrameFormat.RGBA,
+                         CogFrameFormat.BGRx,
+                         CogFrameFormat.BGRx,
+                         CogFrameFormat.ARGB,
+                         CogFrameFormat.xRGB,
+                         CogFrameFormat.ABGR,
+                         CogFrameFormat.xBGR]:
+                self.assertEqual(len(grain.components), 1)
+                self.assertEqual(grain.components[0].stride, 4*width*bps)
+                self.assertEqual(grain.components[0].width, width)
+                self.assertEqual(grain.components[0].height, height)
+                self.assertEqual(grain.components[0].offset, 0)
+                self.assertEqual(grain.components[0].length, width*height*bps*4)
+                self.assertEqual(len(grain.components[0]), 5)
+
+            elif fmt == CogFrameFormat.v216:
+                self.assertEqual(len(grain.components), 1)
+                self.assertEqual(grain.components[0].stride, 2*width*bps)
+                self.assertEqual(grain.components[0].width, width)
+                self.assertEqual(grain.components[0].height, height)
+                self.assertEqual(grain.components[0].offset, 0)
+                self.assertEqual(grain.components[0].length, width*height*bps*2)
+                self.assertEqual(len(grain.components[0]), 5)
+
+            elif fmt == CogFrameFormat.v210:
+                self.assertEqual(len(grain.components), 1)
+                self.assertEqual(grain.components[0].stride, (((width + 47) // 48) * 128))
+                self.assertEqual(grain.components[0].width, width)
+                self.assertEqual(grain.components[0].height, height)
+                self.assertEqual(grain.components[0].offset, 0)
+                self.assertEqual(grain.components[0].length, height*(((width + 47) // 48) * 128))
+                self.assertEqual(len(grain.components[0]), 5)
+
+            else:
+                raise Exception()
+
+            if bps == 1:
+                dtype = np.dtype(np.uint8)
+            elif bps == 2:
+                dtype = np.dtype(np.int16)
+            elif bps == 4:
+                dtype = np.dtype(np.int32)
+            else:
+                raise Exception()
+
+            self.assertIsInstance(grain.data, np.ndarray)
+            self.assertEqual(grain.data.nbytes, grain.expected_length)
+            self.assertEqual(grain.data.dtype, dtype)
+            self.assertEqual(grain.data.size, grain.expected_length//bps)
+            self.assertEqual(grain.data.itemsize, bps)
+            self.assertEqual(grain.data.ndim, 1)
+            self.assertEqual(grain.data.shape, (grain.expected_length//bps,))
+
+            self.assertEqual(repr(grain), "VideoGrain({!r},< numpy data of length {} >)".format(grain.meta, len(grain.data)))
+
+            if fmt == CogFrameFormat.v210:
+                # V210 is barely supported. Convert it to something else to actually use it!
+                self.assertEqual(len(grain.component_data), 0)
+            else:
+                self.assertIsInstance(grain.component_data[0], np.ndarray)
+                self.assertEqual(grain.component_data[0].nbytes, width*height*bps)
+                self.assertEqual(grain.component_data[0].dtype, dtype)
+                self.assertEqual(grain.component_data[0].size, width*height)
+                self.assertEqual(grain.component_data[0].itemsize, bps)
+                self.assertEqual(grain.component_data[0].ndim, 2)
+                self.assertEqual(grain.component_data[0].shape, (width, height))
+
+                self.assertIsInstance(grain.component_data[1], np.ndarray)
+                self.assertEqual(grain.component_data[1].nbytes, width*height*bps >> (hs + vs))
+                self.assertEqual(grain.component_data[1].dtype, dtype)
+                self.assertEqual(grain.component_data[1].size, width*height >> (hs + vs))
+                self.assertEqual(grain.component_data[1].itemsize, bps)
+                self.assertEqual(grain.component_data[1].ndim, 2)
+                self.assertEqual(grain.component_data[1].shape, (width >> hs, height >> vs))
+
+                self.assertIsInstance(grain.component_data[2], np.ndarray)
+                self.assertEqual(grain.component_data[2].nbytes, width*height*bps >> (hs + vs))
+                self.assertEqual(grain.component_data[2].dtype, dtype)
+                self.assertEqual(grain.component_data[2].size, width*height >> (hs + vs))
+                self.assertEqual(grain.component_data[2].itemsize, bps)
+                self.assertEqual(grain.component_data[2].ndim, 2)
+                self.assertEqual(grain.component_data[2].shape, (width >> hs, height >> vs))
+
+                # Test that changes to the component arrays are reflected in the main data array
+                for y in range(0, 16):
+                    for x in range(0, 16):
+                        grain.component_data[0][x, y] = (y*16 + x) & 0x3F
+
+                for y in range(0, 16 >> vs):
+                    for x in range(0, 16 >> hs):
+                        grain.component_data[1][x, y] = (y*16 + x) & 0x3F + 0x40
+                        grain.component_data[2][x, y] = (y*16 + x) & 0x3F + 0x50
+
+                if COG_FRAME_IS_PLANAR(fmt):
+                    for y in range(0, 16):
+                        for x in range(0, 16):
+                            self.assertEqual(grain.data[y*width + x], (y*16 + x) & 0x3F)
+
+                    for y in range(0, 16 >> vs):
+                        for x in range(0, 16 >> hs):
+                            self.assertEqual(grain.data[width*height + y*(width >> hs) + x], (y*16 + x) & 0x3F + 0x40)
+                            self.assertEqual(grain.data[width*height + (width >> hs)*(height >> vs) + y*(width >> hs) + x], (y*16 + x) & 0x3F + 0x50)
+
+                elif fmt in [CogFrameFormat.UYVY, CogFrameFormat.v216]:
+                    for y in range(0, 16):
+                        for x in range(0, 8):
+                            self.assertEqual(grain.data[y*width*2 + 4*x + 0], (y*16 + x) & 0x3F + 0x40)
+                            self.assertEqual(grain.data[y*width*2 + 4*x + 1], (y*16 + 2*x + 0) & 0x3F)
+                            self.assertEqual(grain.data[y*width*2 + 4*x + 2], (y*16 + x) & 0x3F + 0x50)
+                            self.assertEqual(grain.data[y*width*2 + 4*x + 3], (y*16 + 2*x + 1) & 0x3F)
+
+                elif fmt == CogFrameFormat.YUYV:
+                    for y in range(0, 16):
+                        for x in range(0, 8):
+                            self.assertEqual(grain.data[y*width*2 + 4*x + 0], (y*16 + 2*x + 0) & 0x3F)
+                            self.assertEqual(grain.data[y*width*2 + 4*x + 1], (y*16 + x) & 0x3F + 0x40)
+                            self.assertEqual(grain.data[y*width*2 + 4*x + 2], (y*16 + 2*x + 1) & 0x3F)
+                            self.assertEqual(grain.data[y*width*2 + 4*x + 3], (y*16 + x) & 0x3F + 0x50)
+
+                elif fmt == CogFrameFormat.RGB:
+                    for y in range(0, 16):
+                        for x in range(0, 16):
+                            self.assertEqual(grain.data[y*width*3 + 3*x + 0], (y*16 + x) & 0x3F)
+                            self.assertEqual(grain.data[y*width*3 + 3*x + 1], (y*16 + x) & 0x3F + 0x40)
+                            self.assertEqual(grain.data[y*width*3 + 3*x + 2], (y*16 + x) & 0x3F + 0x50)
+
+                elif fmt in [CogFrameFormat.RGBx,
+                             CogFrameFormat.RGBA,
+                             CogFrameFormat.BGRx,
+                             CogFrameFormat.BGRx]:
+                    for y in range(0, 16):
+                        for x in range(0, 16):
+                            self.assertEqual(grain.data[y*width*4 + 4*x + 0], (y*16 + x) & 0x3F)
+                            self.assertEqual(grain.data[y*width*4 + 4*x + 1], (y*16 + x) & 0x3F + 0x40)
+                            self.assertEqual(grain.data[y*width*4 + 4*x + 2], (y*16 + x) & 0x3F + 0x50)
+
+                elif fmt in [CogFrameFormat.ARGB,
+                             CogFrameFormat.xRGB,
+                             CogFrameFormat.ABGR,
+                             CogFrameFormat.xBGR]:
+                    for y in range(0, 16):
+                        for x in range(0, 16):
+                            self.assertEqual(grain.data[y*width*4 + 4*x + 1], (y*16 + x) & 0x3F)
+                            self.assertEqual(grain.data[y*width*4 + 4*x + 2], (y*16 + x) & 0x3F + 0x40)
+                            self.assertEqual(grain.data[y*width*4 + 4*x + 3], (y*16 + x) & 0x3F + 0x50)
+
+                else:
+                    raise Exception()
+
+        return __inner
+
+    def test_video_grain_create(self):
+        src_id = uuid.UUID("f18ee944-0841-11e8-b0b0-17cef04bd429")
+        flow_id = uuid.UUID("f79ce4da-0841-11e8-9a5b-dfedb11bafeb")
+        cts = Timestamp.from_tai_sec_nsec("417798915:0")
+        ots = Timestamp.from_tai_sec_nsec("417798915:5")
+        sts = Timestamp.from_tai_sec_nsec("417798915:10")
+
+        for fmt in [CogFrameFormat.S32_444,
+                    CogFrameFormat.S32_422,
+                    CogFrameFormat.S32_420,
+                    CogFrameFormat.S16_444_10BIT,
+                    CogFrameFormat.S16_422_10BIT,
+                    CogFrameFormat.S16_420_10BIT,
+                    CogFrameFormat.S16_444_12BIT,
+                    CogFrameFormat.S16_422_12BIT,
+                    CogFrameFormat.S16_420_12BIT,
+                    CogFrameFormat.S16_444,
+                    CogFrameFormat.S16_422,
+                    CogFrameFormat.S16_420,
+                    CogFrameFormat.U8_444,
+                    CogFrameFormat.U8_422,
+                    CogFrameFormat.U8_420,
+                    CogFrameFormat.U8_444_RGB,
+                    CogFrameFormat.S16_444_RGB,
+                    CogFrameFormat.S16_444_12BIT_RGB,
+                    CogFrameFormat.S16_444_10BIT_RGB,
+                    CogFrameFormat.UYVY,
+                    CogFrameFormat.YUYV,
+                    CogFrameFormat.RGB,
+                    CogFrameFormat.RGBx,
+                    CogFrameFormat.RGBA,
+                    CogFrameFormat.BGRx,
+                    CogFrameFormat.BGRx,
+                    CogFrameFormat.ARGB,
+                    CogFrameFormat.xRGB,
+                    CogFrameFormat.ABGR,
+                    CogFrameFormat.xBGR,
+                    CogFrameFormat.v216,
+                    CogFrameFormat.v210]:
+            with self.subTest(fmt=fmt):
+                with mock.patch.object(Timestamp, "get_time", return_value=cts):
+                    grain = VideoGrain(src_id, flow_id, origin_timestamp=ots, sync_timestamp=sts,
+                                       cog_frame_format=fmt,
+                                       width=1920, height=1080, cog_frame_layout=CogFrameLayout.FULL_FRAME)
+
+                self.assertIsVideoGrain(fmt)(grain)
+
+                if fmt is not CogFrameFormat.v210:
+                    self.assertComponentsAreModifiable(grain)
+
+    def test_video_grain_create_discontiguous(self):
+        src_id = uuid.UUID("f18ee944-0841-11e8-b0b0-17cef04bd429")
+        flow_id = uuid.UUID("f79ce4da-0841-11e8-9a5b-dfedb11bafeb")
+        cts = Timestamp.from_tai_sec_nsec("417798915:0")
+        ots = Timestamp.from_tai_sec_nsec("417798915:5")
+        sts = Timestamp.from_tai_sec_nsec("417798915:10")
+
+        data = bytearray(11*1024*1024)
+
+        grain = VideoGrain({
+            "grain": {
+                "grain_type": "video",
+                "source_id": src_id,
+                "flow_id": flow_id,
+                "origin_timestamp": ots,
+                "sync_timestamp": sts,
+                "creation_timestamp": cts,
+                "rate": {
+                    "numerator": 25,
+                    "denominator": 1,
+                },
+                "duration": {
+                    "numerator": 1,
+                    "denominator": 25,
+                },
+                "cog_frame": {
+                    "format": CogFrameFormat.S16_422_10BIT,
+                    "width": 1920,
+                    "height": 1080,
+                    "layout": CogFrameLayout.FULL_FRAME,
+                    "extension": 0,
+                    "components": [
+                        {
+                            "stride": 4096,
+                            "width": 1920,
+                            "height": 1080,
+                            "length": 4423680,
+                            "offset": 0
+                        },
+                        {
+                            "stride": 2048,
+                            "width": 960,
+                            "height": 1080,
+                            "length": 2211840,
+                            "offset": 5*1024*1024
+                        },
+                        {
+                            "stride": 2048,
+                            "width": 960,
+                            "height": 1080,
+                            "length": 2211840,
+                            "offset": 8*1024*1024
+                        },
+                    ]
+                }
+            }
+        }, data)
+
+        for y in range(0, 16):
+            for x in range(0, 16):
+                grain.component_data[0][x, y] = (y*16 + x) & 0x3F
+
+        for y in range(0, 16):
+            for x in range(0, 8):
+                grain.component_data[1][x, y] = (y*16 + x) & 0x3F + 0x40
+                grain.component_data[2][x, y] = (y*16 + x) & 0x3F + 0x50
+
+        for y in range(0, 16):
+            for x in range(0, 16):
+                self.assertEqual(grain.data[y*grain.components[0].stride//2 + x], (y*16 + x) & 0x3F)
+
+        for y in range(0, 16):
+            for x in range(0, 8):
+                self.assertEqual(grain.data[grain.components[1].offset//2 + y*grain.components[1].stride//2 + x], (y*16 + x) & 0x3F + 0x40)
+                self.assertEqual(grain.data[grain.components[2].offset//2 + y*grain.components[2].stride//2 + x], (y*16 + x) & 0x3F + 0x50)
+
+    def test_copy(self):
+        src_id = uuid.UUID("f18ee944-0841-11e8-b0b0-17cef04bd429")
+        flow_id = uuid.UUID("f79ce4da-0841-11e8-9a5b-dfedb11bafeb")
+        cts = Timestamp.from_tai_sec_nsec("417798915:0")
+        ots = Timestamp.from_tai_sec_nsec("417798915:5")
+        sts = Timestamp.from_tai_sec_nsec("417798915:10")
+
+        with mock.patch.object(Timestamp, "get_time", return_value=cts):
+            grain = VideoGrain(src_id, flow_id, origin_timestamp=ots, sync_timestamp=sts,
+                               cog_frame_format=CogFrameFormat.S16_422_10BIT,
+                               width=1920, height=1080, cog_frame_layout=CogFrameLayout.FULL_FRAME)
+
+        grain.data[0] = 0x1BBC
+
+        clone = copy(grain)
+
+        self.assertEqual(grain.data[0], clone.data[0])
+
+        grain.data[0] = 0xCAFE
+
+        self.assertEqual(grain.data[0], clone.data[0])
+
+    def test_deepcopy(self):
+        src_id = uuid.UUID("f18ee944-0841-11e8-b0b0-17cef04bd429")
+        flow_id = uuid.UUID("f79ce4da-0841-11e8-9a5b-dfedb11bafeb")
+        cts = Timestamp.from_tai_sec_nsec("417798915:0")
+        ots = Timestamp.from_tai_sec_nsec("417798915:5")
+        sts = Timestamp.from_tai_sec_nsec("417798915:10")
+
+        with mock.patch.object(Timestamp, "get_time", return_value=cts):
+            grain = VideoGrain(src_id, flow_id, origin_timestamp=ots, sync_timestamp=sts,
+                               cog_frame_format=CogFrameFormat.S16_422_10BIT,
+                               width=1920, height=1080, cog_frame_layout=CogFrameLayout.FULL_FRAME)
+
+        grain.data[0] = 0x1BBC
+
+        clone = deepcopy(grain)
+
+        self.assertEqual(grain.data[0], clone.data[0])
+
+        grain.data[0] = 0xCAFE
+
+        self.assertNotEqual(grain.data[0], clone.data[0])